Admin Panel Commit

+++ api/class/advertisement.php
@@ -40,26 +40,28 @@ class Advertisement
         }
     }
     // Create an advertisement OK
-    public function createAdvertisement()
-    {
-        $headers = apache_request_headers();
-        $tokenParts = explode(".", str_replace("Bearer ", "", $headers["Authorization"]));
-        $payload = base64_decode($tokenParts[1]);
-        $signature = hash_hmac("sha256", $tokenParts[0] . "." . $tokenParts[1], "90zgLEniSbKFrV6OJjVa825KcTI1JC7m", true);
-        $base64UrlSignature = str_replace(["+", "/", "="], ["-", "_", ""], base64_encode($signature));
-        if ($base64UrlSignature == $tokenParts[2]) {
-            if (intval(((array) json_decode($payload))["admin"])) {
-                $advertisement = $this->connection->prepare("INSERT INTO advertisements (advertisement_name, advertisement_company, advertisement_location, advertisement_type, advertisement_description, advertisement_salary) VALUES (?, ?, ?, ?, ?, ?)");
-                $advertisement->bindParam(1, htmlspecialchars(strip_tags($this->advertisement_name)));
-                $advertisement->bindParam(2, htmlspecialchars(strip_tags($this->advertisement_company)));
-                $advertisement->bindParam(3, htmlspecialchars(strip_tags($this->advertisement_location)));
-                $advertisement->bindParam(4, htmlspecialchars(strip_tags($this->advertisement_type)));
-                $advertisement->bindParam(5, htmlspecialchars(strip_tags($this->advertisement_description)));
-                $advertisement->bindParam(6, htmlspecialchars(strip_tags($this->advertisement_salary)));
-                if ($advertisement->execute()) {
-                    return array("response" => true);
+    public function createAdvertisement() {
+        if (isset($_COOKIE["token"])) {
+            $tokenParts = explode(".", $_COOKIE["token"]);
+            $payload = base64_decode($tokenParts[1]);
+            $signature = hash_hmac("sha256", $tokenParts[0] . "." . $tokenParts[1], "90zgLEniSbKFrV6OJjVa825KcTI1JC7m", true);
+            $base64UrlSignature = str_replace(["+", "/", "="], ["-", "_", ""], base64_encode($signature));
+            if ($base64UrlSignature == $tokenParts[2]) {
+                if (intval(((array) json_decode($payload))["admin"])) {
+                    $advertisement = $this->connection->prepare("INSERT INTO advertisements (advertisement_name, advertisement_company, advertisement_location, advertisement_type, advertisement_description, advertisement_salary) VALUES (?, ?, ?, ?, ?, ?)");
+                    $advertisement->bindParam(1, htmlspecialchars(strip_tags($this->advertisement_name)));
+                    $advertisement->bindParam(2, htmlspecialchars(strip_tags($this->advertisement_company)));
+                    $advertisement->bindParam(3, htmlspecialchars(strip_tags($this->advertisement_location)));
+                    $advertisement->bindParam(4, htmlspecialchars(strip_tags($this->advertisement_type)));
+                    $advertisement->bindParam(5, htmlspecialchars(strip_tags($this->advertisement_description)));
+                    $advertisement->bindParam(6, htmlspecialchars(strip_tags($this->advertisement_salary)));
+                    if ($advertisement->execute()) {
+                        return array("response" => true);
+                    } else {
+                        return array("response" => false, "access" => true);
+                    }
                 } else {
-                    return array("response" => false, "access" => true);
+                    return array("response" => false, "access" => false);
                 }
             } else {
                 return array("response" => false, "access" => false);
@@ -69,27 +71,29 @@ class Advertisement
         }
     }
     // Update an advertisement OK - when ressource is not found ?
-    public function updateAdvertisement()
-    {
-        $headers = apache_request_headers();
-        $tokenParts = explode(".", str_replace("Bearer ", "", $headers["Authorization"]));
-        $payload = base64_decode($tokenParts[1]);
-        $signature = hash_hmac("sha256", $tokenParts[0] . "." . $tokenParts[1], "90zgLEniSbKFrV6OJjVa825KcTI1JC7m", true);
-        $base64UrlSignature = str_replace(["+", "/", "="], ["-", "_", ""], base64_encode($signature));
-        if ($base64UrlSignature == $tokenParts[2]) {
-            if (intval(((array) json_decode($payload))["admin"])) {
-                $advertisement = $this->connection->prepare("UPDATE advertisements SET advertisement_name = ?, advertisement_company = ?, advertisement_location = ?, advertisement_type = ?, advertisement_description = ?, advertisement_salary = ? WHERE id_advertisement = ?");
-                $advertisement->bindParam(1, htmlspecialchars(strip_tags($this->advertisement_name)));
-                $advertisement->bindParam(2, htmlspecialchars(strip_tags($this->advertisement_company)));
-                $advertisement->bindParam(3, htmlspecialchars(strip_tags($this->advertisement_location)));
-                $advertisement->bindParam(4, htmlspecialchars(strip_tags($this->advertisement_type)));
-                $advertisement->bindParam(5, htmlspecialchars(strip_tags($this->advertisement_description)));
-                $advertisement->bindParam(6, htmlspecialchars(strip_tags($this->advertisement_salary)));
-                $advertisement->bindParam(7, htmlspecialchars(strip_tags($this->id_advertisement)));
-                if ($advertisement->execute()) {
-                    return array("response" => true);
+    public function updateAdvertisement() {
+        if (isset($_COOKIE["token"])) {
+            $tokenParts = explode(".", $_COOKIE["token"]);
+            $payload = base64_decode($tokenParts[1]);
+            $signature = hash_hmac("sha256", $tokenParts[0] . "." . $tokenParts[1], "90zgLEniSbKFrV6OJjVa825KcTI1JC7m", true);
+            $base64UrlSignature = str_replace(["+", "/", "="], ["-", "_", ""], base64_encode($signature));
+            if ($base64UrlSignature == $tokenParts[2]) {
+                if (intval(((array) json_decode($payload))["admin"])) {
+                    $advertisement = $this->connection->prepare("UPDATE advertisements SET advertisement_name = ?, advertisement_company = ?, advertisement_location = ?, advertisement_type = ?, advertisement_description = ?, advertisement_salary = ? WHERE id_advertisement = ?");
+                    $advertisement->bindParam(1, htmlspecialchars(strip_tags($this->advertisement_name)));
+                    $advertisement->bindParam(2, htmlspecialchars(strip_tags($this->advertisement_company)));
+                    $advertisement->bindParam(3, htmlspecialchars(strip_tags($this->advertisement_location)));
+                    $advertisement->bindParam(4, htmlspecialchars(strip_tags($this->advertisement_type)));
+                    $advertisement->bindParam(5, htmlspecialchars(strip_tags($this->advertisement_description)));
+                    $advertisement->bindParam(6, htmlspecialchars(strip_tags($this->advertisement_salary)));
+                    $advertisement->bindParam(7, htmlspecialchars(strip_tags($this->id_advertisement)));
+                    if ($advertisement->execute()) {
+                        return array("response" => true);
+                    } else {
+                        return array("response" => false, "access" => true);
+                    }
                 } else {
-                    return array("response" => false, "access" => true);
+                    return array("response" => false, "access" => false);
                 }
             } else {
                 return array("response" => false, "access" => false);
@@ -99,21 +103,23 @@ class Advertisement
         }
     }
     // Delete an advertisement OK - when ressource is not found ?
-    public function deleteAdvertisement()
-    {
-        $headers = apache_request_headers();
-        $tokenParts = explode(".", str_replace("Bearer ", "", $headers["Authorization"]));
-        $payload = base64_decode($tokenParts[1]);
-        $signature = hash_hmac("sha256", $tokenParts[0] . "." . $tokenParts[1], "90zgLEniSbKFrV6OJjVa825KcTI1JC7m", true);
-        $base64UrlSignature = str_replace(["+", "/", "="], ["-", "_", ""], base64_encode($signature));
-        if ($base64UrlSignature == $tokenParts[2]) {
-            if (intval(((array) json_decode($payload))["admin"])) {
-                $advertisement = $this->connection->prepare("DELETE FROM advertisements WHERE id_advertisement = ?");
-                $advertisement->bindParam(1, htmlspecialchars(strip_tags($this->id_advertisement)));
-                if ($advertisement->execute()) {
-                    return array("response" => true);
+    public function deleteAdvertisement() {
+        if (isset($_COOKIE["token"])) {
+            $tokenParts = explode(".", $_COOKIE["token"]);
+            $payload = base64_decode($tokenParts[1]);
+            $signature = hash_hmac("sha256", $tokenParts[0] . "." . $tokenParts[1], "90zgLEniSbKFrV6OJjVa825KcTI1JC7m", true);
+            $base64UrlSignature = str_replace(["+", "/", "="], ["-", "_", ""], base64_encode($signature));
+            if ($base64UrlSignature == $tokenParts[2]) {
+                if (intval(((array) json_decode($payload))["admin"])) {
+                    $advertisement = $this->connection->prepare("DELETE FROM advertisements WHERE id_advertisement = ?");
+                    $advertisement->bindParam(1, htmlspecialchars(strip_tags($this->id_advertisement)));
+                    if ($advertisement->execute()) {
+                        return array("response" => true);
+                    } else {
+                        return array("response" => false, "access" => true);
+                    }
                 } else {
-                    return array("response" => false, "access" => true);
+                    return array("response" => false, "access" => false);
                 }
             } else {
                 return array("response" => false, "access" => false);
